[FEAT]: JoinupPage 확장을 위한 Handler open class 생성
